Adding ability to use a cell's content for link generation
This change adds an input box to the Link section of the Column Styles
tab in the table panel plugin. This 'Parameterize by' input box takes
a regular expression that will be used to split a cell's content. Once
the cell's content has been split the sub-strings will be available
for URL generation using the $__pattern_n variables starting from
0. The default behavior splits the cell's content by spaces.